# Optimize dwarf line numbers decoding