--- a/src/callstack.cr
+++ b/src/callstack.cr
@@ -285,8 +285,11 @@ struct CallStack
           mach_o.read_section?("__debug_info") do |sh, io|
             names = [] of {LibC::SizeT, LibC::SizeT, String}
 
-            while io.tell - sh.offset < sh.size
-              offset = io.tell - sh.offset
+            while true
+              pos = io.pos
+              offset = pos - sh.offset
+              break unless offset < sh.size
+
               info = Debug::DWARF::Info.new(io, offset)
 
               mach_o.read_section?("__debug_abbrev") do |sh, io|
@@ -387,8 +390,11 @@ struct CallStack
           elf.read_section?(".debug_info") do |sh, io|
             names = [] of {LibC::SizeT, LibC::SizeT, String}
 
-            while io.tell - sh.offset < sh.size
-              offset = io.tell - sh.offset
+            while true
+              pos = io.pos
+              offset = pos - sh.offset
+              break unless offset < sh.size
+
               info = Debug::DWARF::Info.new(io, offset)
 
               elf.read_section?(".debug_abbrev") do |sh, io|