@ysbaddaden 2019-02-11
Impressive boost, especially since the files/directories arrays are usually small if I remember correctly.