@asterite 2019-02-11
@yxhuvud Yes, `find` could use a binary search, I thought of it. That said, I don't understand why `@matrix` is defined like that, maybe pushing all `Row`s in a single array would be enough? Not sure.

But that's not the bottleneck, at least when I profiled it. The bottleneck are:
- `Debug::DWARF::Strings#decode`: most of the time spent it `gets` and `tell`/`pos` (this can be optimized just a bit by using `gets('\0', chomp: true)` but I didn't want to make this PR too big
- `Debug::DWARF.read_unsigned_leb128`: most of the time is spent in `read_byte` 

I'm not sure there's something we can do about those two points. In both cases (I think?) the IO is repositioned and read, and because it's buffered an entire block is consumed. Maybe changing the buffer size could improve this because `decode` usually needs just a few bytes (it's just a name) and I guess the same applies to `read_unsigned_leb128`. But for that we should first make the buffer size of `IO::Buffered` configurable.