@asterite 2019-02-11
@ysbaddaden Yes, the arrays are small. The problem is that the `register_filename` and `register_directory` end up being called, at least in `std_spec`, 1 million times, and most of those calls are always with the same filename/directory that was passed the last time. In fact, another way I tried to optimize this is by only recomputing the file/path indexes when `register.file` was set, or when `Register.new` was done, or at the beginning of `read_statement_program` but the code ended up being a bit more complex and taking more time if I also didn't do the "is it the same as the last time" check.