@asterite 2019-02-12
I added another commit that introduces two temporary hashes that map dir/file-names to indexes, to avoid doing linear scans over `@files` and `@directories`. That reduces the time of `Debug::DWARF::LineNumbers#decode_sequences` from 0.82s to 0.78s.