@asterite 2019-02-12
I just pushed another commit with more improvements. I tried to do small commits for this but I was experimenting and in the end it was a bit hard for me to separate all the changes.

Essentially, I changed some `UInt32 | UInt64` unions to just `UInt64` (the cost of dealing with unions is bigger than just tracking the exact type). Another change is avoiding calling `io.tell` multiple times (this is expensive; I also used `io.pos`, I think we should remove the `tell` alias).

But the biggest change is in `DWARF::Strings`: we now use a small (16Kb) buffer to read strings from instead of seeking and reading from the IO (which is done **a lot** of times).

So now:

```crystal
$ bin/crystal build spec/std_spec.cr
$ ./std_spec -e "inspects with cause"
.

Finished in 1.31 seconds
```

(down from 1.83s)

I'm almost sure more things can be optimized, but with these changes I'm pretty satisfied now.