@wontruefree 2019-02-12
@asterite why use 2 hashes?  Directories and filenames would not collide.  It seems like you could reduce this to one register method and one hash.