@asterite 2019-02-12
@wontruefree `files` and `directories` are public and are used by other parts of the code that need the "index -> name" mapping.