@wontruefree 2019-02-12
@asterite I am not sure I understand.

But if you have considered it and know why it won't work that is cool.  I cannot make a PR into your fork but here is what I was thinking.

https://github.com/crystal-lang/crystal/compare/master...wontruefree:opt/dwarf-decoding