@asterite 2019-02-12
@wontruefree the `@files` and `@directories` variables are exposed through the getters of `LineNumbers` and used [here](https://github.com/crystal-lang/crystal/blob/323ed75fb9edfd99134627b635e1325e605ad8df/src/callstack.cr#L226-L227). In your code you just stop putting data there. Try it with this code:

```crystal
begin
  raise "oh no"
rescue ex
  puts ex.inspect_with_backtrace
end

begin
  [1][2]
rescue ex
  puts ex.inspect_with_backtrace
end
```

 Do you get something correct in the output?