@wontruefree 2019-02-12
@asterite I see what you are saying.
I updated the branch and reduced the external interface to use a method called paths.

After thinking about this why not just make the `index` the instance variable and have the callstack use a hash lookup.  Seems like that would be faster.