@asterite 2019-02-12
I think some tests are allocating too much memory and it dies on Linux 32 bits. But I didn't start investigating that yet.