@asterite 2019-02-13
I just pushed yet another commit. Thanks to what @wontruefree said I noticed that it doesn't make sense to build those `files`/`directories` arrays. Instead, we can store them directly inside each `Row`.

@ysbaddaden Do you see a problem with this?