@asterite 2019-02-13
@ysbaddaden I don't think so because all of the strings are stored inside arrays inside `Sequence`s so a same string shares the same memory for others. But maybe it wasn't like that at some point, I don't know. This way it seems to work :-)

Maybe if I have time I'll learn that format and write specs for it, so we can continue simplifying it (with more confidence). However, right now I think the times are more or less fine so I might focus on something else after this.