@ysbaddaden 2019-02-13
Once the sequences are decoded we don't retain any references to them, so their `@directories` and `@files` arrays will be collected by the GC. I think most folder names and enough file names are shared across sequences that de-duplicating can be significant on memory usage, especially with large programs, but that should be verified.