@Sija 2019-02-13
@asterite Would you mind updating PR description with the newest run-time numbers?