@ysbaddaden 2019-02-15
Let's keep each commit. We could squash "simplify while loop" but that's nitpicking.